+++ .pylintrc
@@ -148,8 +148,7 @@ disable=abstract-method,
         zip-builtin-not-iterating,
         protected-access,
         wrong-import-position,
-        inconsistent-quotes,
-        broad-exception-caught
+        inconsistent-quotes
 
 [REPORTS]
