+++ README.md
@@ -1,9 +1,6 @@
 <div style="text-align: center;">
 
 # SAE-3.05
-[![linting: pylint](https://img.shields.io/badge/linting-pylint-yellowgreen)](https://github.com/pylint-dev/pylint)
-
-[![Coverage Status](https://coveralls.io/repos/github/Yannis-barache/SAE-3.05/badge.svg?branch=develop)](https://coveralls.io/github/Yannis-barache/SAE-3.05?branch=develop)
 
 <div style="display: flex; align-items: center; justify-content:space-evenly">
     <img src="https://assets.onthehub.com/attachments/15/bf1336cd-799b-e011-969d-0030487d8897/40d4e299-03ff-4c99-94a8-7a100f07837c.jpg" alt="Logo IUT Orléans" width="275"/>
