+++ appli/BD/competition_bd.py
@@ -81,16 +81,14 @@ class CompetitionBD:
         :param competition : competition
         """
         try:
-            query = text(
-                f"INSERT INTO COMPETITION (nomCompetition, dateCompetition, "
-                f"dateFinInscription, saisonCompetition,idLieu, idArme, "
-                f"idCategorie, coefficientCompetition) VALUES "
-                f"('{competition.get_nom()}', '{competition.get_date()}', "
-                f"'{competition.get_date_fin_inscription()}', "
-                f"'{competition.get_saison()}', {competition.get_lieu().get_id()}, "
-                f"{competition.get_arme().get_id()}, "
-                f"{competition.get_categorie().get_id()}, {competition.get_coefficient()})"
-            )
+            query = text(f"INSERT INTO COMPETITION (nomCompetition, dateCompetition, "
+                         f"dateFinInscription, saisonCompetition,idLieu, idArme, "
+                         f"idCategorie, coefficientCompetition) VALUES "
+                         f"('{competition.get_nom()}', '{competition.get_date()}', "
+                         f"'{competition.get_date_fin_inscription()}', "
+                         f"'{competition.get_saison()}', {competition.get_lieu()}, "
+                         f"{competition.get_arme()}, "
+                         f"{competition.get_categorie()}, {competition.get_coeficient()})")
             self.__connexion.execute(query)
             self.__connexion.commit()
         except Exception as e:
