+++ appli/BD/escrimeur_bd.py
@@ -88,36 +88,16 @@ class EscrimeurBD:
         :param escrimeur : escrimeur
         """
         try:
-            if escrimeur.get_classement() is None:
-                classement = 'NULL'
-            else:
-                classement = str(escrimeur.get_classement())
-            query = text(
-                f"INSERT INTO ESCRIMEUR (nomEscrimeur, licence, "
-                f"prenomEscrimeur, dateNaissance, nomUtilisateurEscrimeur, "
-                f"mdpEscrimeur, classement, sexeEscrimeur, idClub, idCategorie, "
-                f"arbitrage) VALUES ('{escrimeur.get_nom()}', "
-                f"'{escrimeur.get_licence()}', '{escrimeur.get_prenom()}', "
-                f"'{escrimeur.get_date_naissance()}', "
-                f"'{escrimeur.get_nom_utilisateur()}', "
-                f"'{escrimeur.get_mdp()}', {classement}, "
-                f"'{escrimeur.get_sexe()}', {str(escrimeur.get_club().get_id())}, "
-                f"{str(escrimeur.get_categorie().get_id())}, {str(escrimeur.get_arbitrage())})"
-            )
-            self.__connexion.execute(query)
-            self.__connexion.commit()
-        except Exception as e:
-            print(e)
-            return None
-
-    def delete_escrimeur_by_nom(self, nom: str):
-        """
-        Fonction qui supprime un escrimeur en fonction de son nom
-        :param nom: nom de l'escrimeur
-        """
-        try:
-            query = text("DELETE FROM ESCRIMEUR WHERE nomEscrimeur ='" + nom +
-                         "'")
+            query = text(f"INSERT INTO ESCRIMEUR (nomEscrimeur, licence, "
+                         f"prenomEscrimeur, dateNaissance, nomUtilisateurEscrimeur, "
+                         f"mdpEscrimeur, classement, sexeEscrimeur, idClub, idCategorie, "
+                         f"arbitrage) VALUES ('{escrimeur.get_nom()}', "
+                         f"'{escrimeur.get_licence()}', '{escrimeur.get_prenom()}', "
+                         f"'{escrimeur.get_date_naissance()}', "
+                         f"'{escrimeur.get_nom_utilisateur()}', "
+                         f"'{escrimeur.get_mdp()}', {str(escrimeur.get_classement())}, "
+                         f"'{escrimeur.get_sexe()}', {str(escrimeur.get_club())}, "
+                         f"{str(escrimeur.get_categorie())}, {str(escrimeur.get_arbitrage())})")
             self.__connexion.execute(query)
             self.__connexion.commit()
         except Exception as e:
