+++ appli/BD/poule_bd.py
@@ -36,28 +36,13 @@ class PouleBD:
             print(e)
             return None
 
-    def insert_poule(self, une_poule: Poule):
+    def insert_poule(self):
         """
         Fonction qui insère une poule
         :param poule: poule
         """
         try:
-            query = text("INSERT INTO POULE (idPoule) VALUES (" +
-                         str(une_poule.get_id()) + ")")
-            self.__connexion.execute(query)
-            self.__connexion.commit()
-        except Exception as e:
-            print(e)
-            return None
-
-    def delete_poule(self, une_poule: Poule):
-        """
-        Fonction qui supprime une poule
-        :param poule: poule
-        """
-        try:
-            query = text("DELETE FROM POULE WHERE idPoule = " +
-                         str(une_poule.get_id()))
+            query = text("INSERT INTO POULE DEFAULT VALUES")
             self.__connexion.execute(query)
             self.__connexion.commit()
         except Exception as e:
