+++ appli/modele/poule.py
@@ -493,75 +493,5 @@ class Poule:
         indice = touche_marquee - touche_prise
         return (-victoires, -indice)  # Tri décroissant
 
-    def get_nb_victoires(self, escrimeur: Escrimeur) -> int:
-        """
-        Fonction qui retourne le nombre de victoires d'un escrimeur
-
-        Args:
-            escrimeur (Escrimeur): l'escrimeur
-
-        Returns:
-            int: le nombre de victoires de l'escrimeur
-        """
-        cpt = 0
-        for match in self.__les_matchs:
-            if match.get_gagnant() == escrimeur:
-                cpt += 1
-        return cpt
-
-    def get_nb_escrimeurs(self) -> int:
-        """
-        Fonction qui retourne le nombre d'escrimeurs de la poule
-
-        Returns:
-            int: le nombre d'escrimeurs de la poule
-        """
-        return len(self.__les_escrimeurs)
-
-    def get_nb_matchs(self) -> int:
-        """
-        Fonction qui retourne le nombre de matchs de la poule
-
-        Returns:
-            int: le nombre de matchs de la poule
-        """
-        return len(self.__les_matchs)
-
-    def get_nb_touche_marquee(self, escrimeur: Escrimeur) -> int:
-        """
-        Fonction qui retourne le nombre de touche marquée par un escrimeur
-
-        Args:
-            escrimeur (Escrimeur): l'escrimeur
-
-        Returns:
-            int: le nombre de touche marquée par l'escrimeur
-        """
-        cpt = 0
-        for match in self.__les_matchs:
-            if match.get_escrimeur1() == escrimeur:
-                cpt += match.get_nb_touche(escrimeur)
-            elif match.get_escrimeur2() == escrimeur:
-                cpt += match.get_nb_touche(escrimeur)
-        return cpt
-
-    def get_nb_touche_prise(self, escrimeur: Escrimeur) -> int:
-        """
-        Fonction qui retourne le nombre de touche prise par un escrimeur
-
-        Args:
-            escrimeur (Escrimeur): l'escrimeur
-
-        Returns:
-            int: le nombre de touche prise par l'escrimeur
-        """
-        cpt = 0
-        for match in self.__les_matchs:
-            if match.get_escrimeur1() == escrimeur:
-                cpt += match.get_nb_touche(match.get_escrimeur2())
-            elif match.get_escrimeur2() == escrimeur:
-                cpt += match.get_nb_touche(match.get_escrimeur1())
-        return cpt
-
     def __str__(self):
         return f'Poule : {self.__id} |'
