+++ tests/tests_modele/test_competition.py
@@ -15,9 +15,6 @@ from arme import Arme
 from lieu import Lieu
 from escrimeur import Escrimeur
 from exceptions import PasAssezDArbitres
-from club import Club
-from touche import Touche
-from phase_final import PhaseFinal
 
 
 class TestCompetition(unittest.TestCase):
@@ -335,151 +332,6 @@ class TestCompetition(unittest.TestCase):
         assert Competition.nombre_poule(11, 2) == (2, 6)
         with self.assertRaises(PasAssezDArbitres):
             Competition.nombre_poule(10, 1)
-
-    def test_etablir_classement_provisoire(self):
-        """
-        Test de la fonction etablir_classement_provisoire de la classe
-        """
-        club = Club(1, 'nom', 'adresse', 'pays')
-        club2 = Club(2, 'nom2', 'adresse2', 'pays2')
-        escrimeur1 = Escrimeur(1, 'nom', 'prenom', 'sexe', 'date_naissance',
-                               'pseudo', 'mdp', 'licence', 1, club, None,
-                               False)
-        escrimeur2 = Escrimeur(2, 'nom2', 'prenom2', 'sexe2',
-                               'date_naissance2', 'pseudo2', 'mdp2',
-                               'licence2', 2, club, None, False)
-        escrimeur3 = Escrimeur(3, 'nom3', 'prenom3', 'sexe3',
-                               'date_naissance3', 'pseudo3', 'mdp3',
-                               'licence3', 3, club, None, False)
-        escrimeur4 = Escrimeur(4, 'nom4', 'prenom4', 'sexe4',
-                               'date_naissance4', 'pseudo4', 'mdp4',
-                               'licence4', 4, club, None, False)
-        escrimeur5 = Escrimeur(5, 'nom5', 'prenom5', 'sexe5',
-                               'date_naissance5', 'pseudo5', 'mdp5',
-                               'licence5', 5, club, None, False)
-        escrimeur6 = Escrimeur(6, 'nom6', 'prenom6', 'sexe6',
-                               'date_naissance6', 'pseudo6', 'mdp6',
-                               'licence6', 6, club2, None, False)
-        escrimeur7 = Escrimeur(7, 'nom7', 'prenom7', 'sexe7',
-                               'date_naissance7', 'pseudo7', 'mdp7',
-                               'licence7', 7, club2, None, False)
-        escrimeur8 = Escrimeur(8, 'nom8', 'prenom8', 'sexe8',
-                               'date_naissance8', 'pseudo8', 'mdp8',
-                               'licence8', 8, club2, None, False)
-        escrimeur9 = Escrimeur(9, 'nom9', 'prenom9', 'sexe9',
-                               'date_naissance9', 'pseudo9', 'mdp9',
-                               'licence9', 9, club2, None, False)
-        escrimeur10 = Escrimeur(10, 'nom10', 'prenom10', 'sexe10',
-                                'date_naissance10', 'pseudo10', 'mdp10',
-                                'licence10', 10, club, None, False)
-
-        arbitre1 = Escrimeur(11, 'nom11', 'prenom11', 'sexe11',
-                             'date_naissance11', 'pseudo10', 'mdp11',
-                             'licence11', 11, club, None, True)
-        arbitre2 = Escrimeur(12, 'nom12', 'prenom12', 'sexe12',
-                             'date_naissance12', 'pseudo10', 'mdp12',
-                             'licence12', 12, club2, None, True)
-
-        les_escrimeurs = [
-            escrimeur1, escrimeur2, escrimeur3, escrimeur4, escrimeur5,
-            escrimeur6, escrimeur7, escrimeur8, escrimeur9, escrimeur10
-        ]
-        les_arbitres = [arbitre1, arbitre2]
-
-        poules = Competition.generation_poule(les_escrimeurs, les_arbitres)
-        for poule in poules:
-            matchs = poule.generer_matchs(poules.get(poule), 10.5)
-            liste = []
-            for match in matchs:
-                escrimeur11 = match.get_escrimeur1()
-                escrimeur22 = match.get_escrimeur2()
-                touche = Touche(match, escrimeur11, 1)
-                touche2 = Touche(match, escrimeur22, 2)
-                touche3 = Touche(match, escrimeur11, 3)
-                match.ajouter_touche(touche)
-                match.ajouter_touche(touche2)
-                match.ajouter_touche(touche3)
-                match.set_finis(True)
-                liste.append(match)
-
-        classement = Competition.etablir_classement_provisoire(poules)
-        self.assertEqual(len(classement), 10)
-
-    def test_get_puissance_sup(self):
-        """
-        Test de la fonction get_puissance_sup de la classe Competition
-        """
-        self.assertEqual(Competition.get_puissance_sup(10), 4)
-
-    def test_generer_phase_finale(self):
-        """
-        Test de la fonction generer_phase_finale de la classe Competition
-        """
-        club = Club(1, 'nom', 'adresse', 'pays')
-        club2 = Club(2, 'nom2', 'adresse2', 'pays2')
-        escrimeur1 = Escrimeur(1, 'nom', 'prenom', 'sexe', 'date_naissance',
-                               'pseudo', 'mdp', 'licence', 1, club, None,
-                               False)
-        escrimeur2 = Escrimeur(2, 'nom2', 'prenom2', 'sexe2',
-                               'date_naissance2', 'pseudo2', 'mdp2',
-                               'licence2', 2, club, None, False)
-        escrimeur3 = Escrimeur(3, 'nom3', 'prenom3', 'sexe3',
-                               'date_naissance3', 'pseudo3', 'mdp3',
-                               'licence3', 3, club, None, False)
-        escrimeur4 = Escrimeur(4, 'nom4', 'prenom4', 'sexe4',
-                               'date_naissance4', 'pseudo4', 'mdp4',
-                               'licence4', 4, club, None, False)
-        escrimeur5 = Escrimeur(5, 'nom5', 'prenom5', 'sexe5',
-                               'date_naissance5', 'pseudo5', 'mdp5',
-                               'licence5', 5, club, None, False)
-        escrimeur6 = Escrimeur(6, 'nom6', 'prenom6', 'sexe6',
-                               'date_naissance6', 'pseudo6', 'mdp6',
-                               'licence6', 6, club2, None, False)
-        escrimeur7 = Escrimeur(7, 'nom7', 'prenom7', 'sexe7',
-                               'date_naissance7', 'pseudo7', 'mdp7',
-                               'licence7', 7, club2, None, False)
-        escrimeur8 = Escrimeur(8, 'nom8', 'prenom8', 'sexe8',
-                               'date_naissance8', 'pseudo8', 'mdp8',
-                               'licence8', 8, club2, None, False)
-        escrimeur9 = Escrimeur(9, 'nom9', 'prenom9', 'sexe9',
-                               'date_naissance9', 'pseudo9', 'mdp9',
-                               'licence9', 9, club2, None, False)
-        escrimeur10 = Escrimeur(10, 'nom10', 'prenom10', 'sexe10',
-                                'date_naissance10', 'pseudo10', 'mdp10',
-                                'licence10', 10, club, None, False)
-
-        arbitre1 = Escrimeur(11, 'nom11', 'prenom11', 'sexe11',
-                             'date_naissance11', 'pseudo10', 'mdp11',
-                             'licence11', 11, club, None, True)
-        arbitre2 = Escrimeur(12, 'nom12', 'prenom12', 'sexe12',
-                             'date_naissance12', 'pseudo10', 'mdp12',
-                             'licence12', 12, club2, None, True)
-
-        les_escrimeurs = [
-            escrimeur1, escrimeur2, escrimeur3, escrimeur4, escrimeur5,
-            escrimeur6, escrimeur7, escrimeur8, escrimeur9, escrimeur10
-        ]
-        les_arbitres = [arbitre1, arbitre2]
-
-        poules = Competition.generation_poule(les_escrimeurs, les_arbitres)
-        for poule in poules:
-            matchs = poule.generer_matchs(poules.get(poule), 10.5)
-            liste = []
-            for match in matchs:
-                escrimeur11 = match.get_escrimeur1()
-                escrimeur22 = match.get_escrimeur2()
-                touche = Touche(match, escrimeur11, 1)
-                touche2 = Touche(match, escrimeur22, 2)
-                touche3 = Touche(match, escrimeur11, 3)
-                match.ajouter_touche(touche)
-                match.ajouter_touche(touche2)
-                match.ajouter_touche(touche3)
-                match.set_finis(True)
-                liste.append(match)
-        phase, matchs_finaux = Competition.generer_phase_finale(
-            poules, les_arbitres, 10.3)
-        self.assertEqual(len(matchs_finaux), 8)
-        self.assertIsInstance(phase, PhaseFinal)
 
     def test_str(self):
         """
